Update xspectrum1d.py
Minor fix to make nchunks integer

+++ linetools/spectra/xspectrum1d.py
@@ -1433,7 +1433,7 @@ class XSpectrum1D(object):
         else:
             if edges is None:
                 nchunks = max(3, (wmax - wmin) / float(dw))
-                edges = np.linspace(wmin, wmax, nchunks + 1)
+                edges = np.linspace(wmin, wmax, np.int(nchunks) + 1)
 
         if knots is None:
             knots, indices, masked = prepare_knots(
